Fix nullable integer replace in match style calculations

+++ utils/utils_warnings.py
@@ -864,7 +864,13 @@ def calculate_match_style_score_per_match(df):
     # Výpočty
     df["Tempo"] = df["HS"] + df["AS"] + df["HC"] + df["AC"] + df["HF"] + df["AF"]
     df["Goly"] = df["FTHG"] + df["FTAG"]
-    shots_on_target = (df["HST"] + df["AST"]).astype(float).replace(0, 0.1)
+    # Pandas' nullable integer dtype causes `replace` to reject floating
+    # values when performed directly on the summed column.  Cast the input
+    # columns to float before addition so the result is a regular float Series
+    # and zeros can safely be replaced with a small non-zero value.
+    shots_on_target = (
+        df["HST"].astype(float) + df["AST"].astype(float)
+    ).replace(0, 0.1)
     df["Konverze"] = (df["FTHG"] + df["FTAG"]) / shots_on_target
     df["Agrese"] = df["HY"] + df["AY"] + 2 * (df["HR"] + df["AR"]) + df["HF"] + df["AF"]
